remove TryConvertArgument<T>, add Argument<T> ctor with name and ParseArgument<T>
