+++ src/System.CommandLine/Argument{T}.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.CommandLine.Binding;
 using System.CommandLine.Parsing;
 
 namespace System.CommandLine
@@ -38,35 +37,16 @@ namespace System.CommandLine
             SetDefaultValueFactory(() => getDefaultValue());
         }
 
-        public Argument(TryConvertArgument<T> convert, Func<T> getDefaultValue = default) : this()
+        public Argument(
+            string name,
+            ParseArgument<T> parse,
+            bool isDefault = false) : this()
         {
-            if (convert == null)
+            if (!string.IsNullOrWhiteSpace(name))
             {
-                throw new ArgumentNullException(nameof(convert));
+                Name = name;
             }
 
-            ConvertArguments = (ArgumentResult result, out object value) =>
-            {
-                if (convert(result, out var valueObj))
-                {
-                    value = valueObj;
-                    return true;
-                }
-                else
-                {
-                    value = default;
-                    return false;
-                }
-            };
-
-            if (getDefaultValue != default)
-            {
-                SetDefaultValueFactory(() => getDefaultValue());
-            }
-        }
-
-        public Argument(ParseArgument<T> parse, bool isDefault = false) : this()
-        {
             if (parse == null)
             {
                 throw new ArgumentNullException(nameof(parse));
@@ -80,7 +60,7 @@ namespace System.CommandLine
             ConvertArguments = (ArgumentResult argumentResult, out object value) =>
             {
                 var result = parse(argumentResult);
-                
+
                 if (string.IsNullOrEmpty(argumentResult.ErrorMessage))
                 {
                     value = result;
@@ -92,6 +72,10 @@ namespace System.CommandLine
                     return false;
                 }
             };
+        }
+
+        public Argument(ParseArgument<T> parse, bool isDefault = false) : this(null, parse, isDefault)
+        {
         }
     }
 }
