+++ src/System.CommandLine/Binding/TryConvertArgument.cs
@@ -8,16 +8,4 @@ namespace System.CommandLine.Binding
     internal delegate bool TryConvertArgument(
         ArgumentResult argumentResult,
         out object value);
-
-    /// <summary>
-    /// Converts an <see cref="ArgumentResult"/> into an instance of <typeparamref name="T"/>.
-    /// </summary>
-    /// <typeparam name="T">The type to convert to.</typeparam>
-    /// <param name="argumentResult">The <see cref="ArgumentResult"/> representing parsed input to be converted.</param>
-    /// <param name="value">The converted result.</param>
-    /// <returns>True if the conversion is successful; otherwise, false.</returns>
-    /// <remarks>Validation errors can be returned by setting <see cref="SymbolResult.ErrorMessage"/>.</remarks>
-    public delegate bool TryConvertArgument<T>(
-        ArgumentResult argumentResult,
-        out T value);
 }
